use std::size_t instead of std::uint_least128_t for contexnt type

+++ include/fast_io_freestanding_impl/serializations/leb128.h
@@ -132,7 +132,7 @@ inline constexpr char_type* print_reserve_define(
 
 struct leb128_scan_state_t
 {
-	std::uint_least64_t group_count{};
+	std::size_t group_count{};
 };
 
 template <std::integral char_type, typename I>
@@ -146,7 +146,7 @@ namespace details
 {
 template <std::integral char_type, ::fast_io::details::my_integral I>
 inline constexpr parse_result<char_type const*> scn_ctx_define_leb128_impl(
-	std::uint_least64_t& group_count,
+	std::size_t& group_count,
 	char_type const* begin, char_type const* end,
 	I& t) noexcept
 {
@@ -155,7 +155,7 @@ inline constexpr parse_result<char_type const*> scn_ctx_define_leb128_impl(
 	constexpr auto digits{ std::numeric_limits<U>::digits };
 	constexpr auto remains{ digits % 7 };
 	U tmp{};
-	std::uint_least64_t cnt{ group_count };
+	std::size_t cnt{ group_count };
 	if (cnt == 0)
 		t = 0;
 	for (; begin != end; cnt += 7)
@@ -201,7 +201,7 @@ inline constexpr parse_result<char_type const*>
 	constexpr auto digits{ std::numeric_limits<U>::digits };
 	constexpr auto remains{ digits % 7 };
 	U tmp{};
-	std::uint_least64_t cnt{};
+	std::size_t cnt{};
 	t = 0;
 	for (; begin != end; cnt += 7)
 	{
